This commit was manufactured by cvs2svn to create tag 'v1_24_0_release'.

+++ src/Packages/BioPSE/Dataflow/Modules/Forward/DipoleInSphere.cc
@@ -161,7 +161,7 @@ void DipoleInSphere::execute() {
 	}
       }
       
-      DenseMatrix dip_mtrx(pos.size(), 6);
+      DenseMatrix dip_mtrx((int)pos.size(), 6);
       unsigned int i;
       msgStream_ << "Dipoles: " << endl;
       for (i=0; i<pos.size(); ++i){
